Only disable indicator outside of output mode
